moved textstripper to its own file
